[graphite] Break DrawTypeFlags::kText into finer grained draw types
At least for the Chrome pipeline cases, the current kText draw type vastly over-generates pipelines (they just want BitmapTextRenderStep[mask] draws). This CL gives clients greater control over the RenderSteps being used.

This CL also improves the draw testing in the PaintParamsKeys test. Before, we were only exercising plain Bitmap mask text drawing.

Bug: b/349144902
Bug: b/358074434
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/888796
Commit-Queue: Robert Phillips <robertphillips@google.com>

+++ include/gpu/graphite/precompile/PaintOptions.h
@@ -25,7 +25,7 @@ class PrecompileMaskFilter;
 class PrecompileShader;
 
 enum class Coverage;
-enum DrawTypeFlags : uint8_t;
+enum DrawTypeFlags : uint16_t;
 enum class PrecompileImageFilterFlags : uint32_t;
 
 class KeyContext;
